factor out section
